Fix serialization fallbacks, fix lots of things, do stuff, sorry, too many things in this commit. My Bad

+++ C2GUILauncher/src/FileHelpers.cs
@@ -40,8 +40,11 @@ namespace C2GUILauncher {
                 return false;
             }
 
-            logger.Info("Deleting file: " + filePath);
-            File.Delete(filePath);
+            if (File.Exists(filePath)) {
+                logger.Info("Deleting file: " + filePath);
+                File.Delete(filePath);
+            }
+
             return true;
         }
